Refactor repo session helper boundaries
Move repo session search, RPC typing, and shared session result types into focused modules so future changes stay local and easier to review.

Made-with: Cursor
